\name{defectPrediction-package}
\alias{defectPrediction-package}
\alias{defectPrediction}
\docType{package}
\title{
Predict software defects using a time series model.
}
\description{
For predicting defects in a software project using a time series modelling approach, given historical data from the project issue tracking system. A vector AR model with exogenous variables (VARX) model is used. The model structure has number of defect (created) as an endogenous variable, and number of new features (resolved) and number of improvements (resolved) as exogenous variables.
}
\details{
\tabular{ll}{
Package: \tab defectPrediction\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-05\cr
License: \tab Unlimited\cr
}
}
\author{
James Tunnell

Maintainer: James Tunnell <jamestunnell@gmail.com>
}