% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defectPrediction.R
\name{model.regime}
\alias{model.regime}
\title{Time series modeling methodology over a sliding window.}
\usage{
model.regime(ts.data, window.size, ndiff = 0, normality.signif = 0.1,
  conf.levels = c(75, 90), out.dir = NULL, verbose = FALSE, K.min = 4,
  write.extra = F)
}
\arguments{
\item{ts.data}{An xts time series}

\item{window.size}{The number of samples to include in a window for modeling.}

\item{ndiff}{The number of differences to take. Leave at 0 if no differencing is needed.}

\item{conf.levels}{Confidence level(s) to use in testing model forecast performance.}

\item{verbose}{If TRUE, extra info is printed}

\item{K.min}{The minimum allowable ratio of observations to model parameters
Should be greater than 0 and less than 100.}

\item{write.extra}{If TRUE, write extra output files for all windows (one-step predictions and forecasting of hypotheticals)}

\item{norm.signif}{Significance (alpha level) to use in normality testing of model residuals}

\item{our.dir}{Path to a directory where plots can be saved as files}
}
\value{
The value returned is a list
\item{n.windows}{The total number of sample windows where methodology was applied.}
\item{n.nonevalid}{The number of windows where no valid model was found.}
\item{n.nonnormal}{The number of windows where model residuals were non-normal.}
\item{n.outconf}{The number of windows where the model forecast were outside
confidence level(s). This is a named vector.}
\item{n.inconf}{The number of windows where the model forecast were within
confidence level(s). This is a named vector.}
\item{fc.errs}{Forecast errors from all the sample windows}
\item{model.orders}{The model orders used in all windows}
}
\description{
A modeling regime for setting up software defect prediction model, built using
issue tracking system data. Performs the modeling methodology repeatedly over
a sliding sample window.
}
\examples{
\dontrun{
ts.data <- pre.modeling("~/issues.txt", 7, start.date = "2002-01-01",
 out.dir="~/testrun")
result <- model.regime(ts.data, 48, ndiff = 1, out.dir="~/testrun")
}
}
\author{
James Tunnell
}

