% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defectPrediction.R
\name{pre.modeling}
\alias{pre.modeling}
\title{Pre-modeling: issue loading, sampling, and stationarity testing.}
\usage{
pre.modeling(issues.file, sampling.period, max.ndiff = 2, start.date = NULL,
  end.date = NULL, out.dir = NULL)
}
\arguments{
\item{issues.file}{A text file, containing CSV-like table, with software issue data.}

\item{sampling.period}{Sampling period length, in days.}

\item{max.ndiff}{Maximum number of differences to take for stationarity evaluation (and plotting)}

\item{start.date}{The date to start sampling time series data. If left NULL, then sampling begins as early as possible.}

\item{end.date}{The date to stop sampling time series data. If left NULL, then sampling ends as late as possible.}

\item{our.dir}{Path to a directory where stationarity report(s) can be saved as files}
}
\value{
The value returned is a list:
\item{ts}{An xts time series with data sampled from issues}
\item{needs.diffed}{If time series are all stationary, this will be FALSE.
Otherwise, it will be TRUE.}
}
\description{
Pre-modeling: issue loading, sampling, and stationarity testing.
}
\note{
The issues file that is being used should be a text table, and loadable by
\code{read.table(issues.file, header = T)} alone. The table should have
columns for: created, resolved, and type. The type must be one of: "bug",
"improvement", or "newfeature". The created and resolved fields are for
date-time string that can be interpreted using \code{timeDate(mystr)}.
}
\examples{
\dontrun{
ts.data <- pre.modeling("~/issues.txt", 7, start.date = "2002-01-01",
 out.dir="~/testrun")
}
}
\author{
James Tunnell
}

